Improvements (#54)
* Update queries.

* Wrap API error with response class and data result with Either class.

* Enable debugging.

* Create API response wrapper class.

* Fix episode images not loading and cleanup season and episode queries

* Set hardwareAccelerated to true

* Wrap API calls with ApiResponse class

* Update queries.

* Cleanup: Update repositories, cleanup mappers and mockdata, remove unused functions.

* Add generic exception.

* Update states.

* Update TopBarView and add animation

* Minor cleanup

* Minor cleanup

* Update detailView. Fix stretch down gesture and add toolbar title animation.

* Minor cleanup.

* Delete test module.

* Switch to runTest and cleanup tests.

* Spilt up trakt repository and to profile and show repository.

* Use correct repository after splitting it up.

* Update query

* Update error component.

* Update states.

* Remove enableFeaturePreview

* Update gradle build and test commands.

* Update gradle wrapper version.

* Disable hierarchicalStructureSupport

* Build src migration: Herding Elephants. (#53)

* Get rid of buildSrc and pre-compiled scripts

* create gradle plugins.

* Switch to plugin scripts and cleanup dependencies.

* Update integers to long

* Update import after upgrading sqlDelight

* Create kmm scripts

* Provide coroutineContext

* Provide coroutineContext

* Use kmm plugins & update dependencies.

* Delete unused class

* Cleanup: Move common block to script

* Cleanup: update gradle scripts.

* Minor cleanup: add manifestPlaceholders, add suppress annotation.

* Add app flavor extension.

* Simplify build job

* Remove unused blank sourceSets

* Remove unused dependency

* Update lib

* Add flow preview to opt-in.

* Update dependencies.

* Add artwork

* Update readme.

* Propagate error downstream

* Catch exception.

* Remove network exception and rename class.

* Add helper function and optIn annotations.

* Minor cleanup

* Add optIn annotation.

* Update gradle wrapper version.

* Combine show detail requests into one call and update UI states.

* Fix failing test.

* Move modules from domain to data module.

* Refactor and rename show module to category module:

- Move show related implementation classes to tmdb and trakt module.

* Remove database dependency.

* update dependencies.

* Add domain plugin.

* Split up module to data and domain.

- Statemachine implementation has been moved to domain module. This might be renamed to presentation later.

* Delete shows module.

* Update imports.

* Update module dependencies.

* Use expect actual pattern to add koin injection support.

* Use expect actual pattern to add koin injection support.

* Update imports and add data and domain modules.

* Rename package name

* Update setup instructions: Use JDK 17 instead of 11

* Minor cleanup: Rename class and update imports.

* Update modules.

* Load individual items independently preventing blocking the UI

* Delete unused classes.

* Rename episode table id column

* Rename function

* Add functions to save and get season id.

* Some minor cleanup.

* Update JDK to 17 and cancel jobs is there's a new one

* Change EpisodeArtWork return type to flow. This fixes episode images not being updated.

* Add coroutine parameter.

* Use datastore to save and retrieve the seasonId.

* Update dependencies.

* Minor cleanup: Change class to interface

* Minor cleanup: Get rid of mutableStateFlow and move code blocks to functions.

* Add error test case and fix failing test.

* Set minimum height to episode item

+++ android/features/season-details/src/main/java/com/thomaskioko/tvmaniac/seasondetails/SeasonEpisodeList.kt
@@ -32,13 +32,14 @@ import com.thomaskioko.tvmaniac.compose.components.ColumnSpacer
 import com.thomaskioko.tvmaniac.compose.theme.TvManiacTheme
 import com.thomaskioko.tvmaniac.compose.theme.green
 import com.thomaskioko.tvmaniac.compose.theme.yellow300
+import com.thomaskioko.tvmaniac.data.seasondetails.model.SeasonDetails
 import com.thomaskioko.tvmaniac.resources.R
-import com.thomaskioko.tvmaniac.seasondetails.api.model.SeasonDetails
+import com.thomaskioko.tvmaniac.seasondetails.components.EpisodeListItem
 
 @Composable
 fun SeasonEpisodeList(
     season: SeasonDetails,
-    onEpisodeClicked: (Int) -> Unit = {},
+    onEpisodeClicked: (Long) -> Unit = {},
     index: Int,
     collapsedState: SnapshotStateList<Boolean>
 ) {
